Backend/Netra_Backend/requirements.txt
openc3
pika
psycopg2-binary
